Set child workspace to '.'
By default, with matrix projects, the git repo will be checked
out in $DIR, and then copied to $DIR/systems/$HOSTNAME. This
is a massive waste of disk space for large git checkouts. This
behaviour makes sense when the matrix is built from other
axes, but when the only axis is the hostname, the copying is
not required. Setting child workspace to '.' stops this copying

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>