guests: Implement flavors
Our CI infrastructure and developers have different requirements,
but really the overlap is almost complete and it's a shame that
we require developers to perform manual steps before we can use
our tools.

Flavors are a very simple and effective way to deal with the
issue: we'll be able to configure guests differently based on
whether they will be used for CI or testing.

The default flavor is 'test', which doesn't require the vault
password and as such can be used by anyone out of the box: the
Jenkins setup is skipped in this case. The previous behavior can
be obtained by using the 'ci' flavor instead.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Pavel Hrdina <phrdina@redhat.com>