guests: Use non-debug kernel on Fedora Rawhide
By default, Fedora Rawhide uses a kernel with a lot of debug options
enable. This is very useful for the development of Rawhide itself,
but also causes a pretty significant performance penalty.

Enable the fedora-rawhide-kernel-nodebug, with higher than default
priority, to make sure the non-debug kernels are picked up instead
of the debug ones.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Pavel Hrdina <phrdina@redhat.com>