guests: Rename 'ci' flavor to 'jenkins'
Having the name of the user that can access the machine match
the name of the flavor, as is already the case for the 'test'
flavor, makes some things easier.

We can change this safely because the CentOS CI is the only
user of the 'ci' flavor, and we can easily coordinate a
configuration change to avoid breakages.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>