guests: Configure Jenkins agent based on secret availability
We're going to remove the 'jenkins' pseudo-project from the list
of per-guest projects soon, so we need another way of deciding
whether to install and configure the Jenkins agent.

The availability of the Jenkins secret in the vault is a perfect
candidate, and using it improves things in general because we can
now store the information about which guests are part of the
Jenkins setup in a single place instead of duplicating it.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Pavel Hrdina <phrdina@redhat.com>