guests: Remove ad-hoc package installation
There is a small number of packages that we install as part of
the 'base' task with an ad-hoc call to the package module.

Since we have generic facilities for installing packages and a
generic 'base' pseudo-project that we use for all packages that
are not related to any specific project, we can fold everything
into it and have a single source of truth.

The change requires us to delay user creation, though, because
as part of that we edit the sudoers file, which doesn't exist
until the sudo package has been installed.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Pavel Hrdina <phrdina@redhat.com>