guests: Add Python configuration to shell profile
This is conceptually identical to how Perl is handled, except
that we still need to override $PYTHONPATH in Jenkins jobs in
order to be able to build projects using both Python 3 and
Python 2 on the same guest. We might want to reconsider that
decision, but for the time being this change is still helpful
to people running guests outside of the CI environment.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Pavel Hrdina <phrdina@redhat.com>