jobs: Build using $PYTHON
The current setup is fairly complicated and doesn't buy us
much in terms of coverage: we can default to Python 3 wherever
it's available, and still test Python 2 builds pass thanks to
CentOS, which doesn't have Python 3 yet, and the RPM build jobs,
which build both variants if at all possible.

This reverts commit 57980d405d631f4ce2bfffcec343e506fde343f9
and then some.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Pavel Hrdina <phrdina@redhat.com>