jobs: Introduce machine groups
We're running most of the jobs on all machines, with the major
notable exceptions being the various *-rpm jobs, which of course
only make sense for RPM-based distributions, and the MinGW jobs,
which we only run on Fedora Rawhide.

Instead of listing machines over and over again, define two
list globally and refer to them by name. Ad-hoc machine lists
are still needed in a few places, but overall this cuts down on
repetition significantly.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>