jobs: Unset $CC for Go and MinGW jobs
We want $CC to be set so that ccache is used by default, but
some jobs (Go and MinGW) don't like that very much.

Later on, we might consider building a link farm so that we
can use ccache without having to set $CC; for the time being,
work around the issue at the job level.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>