guests: Force use of nodebug kernel on Rawhide
The nodebug kernel provides much better performance than the
standard Rawhide kernel, so we want to use it; however, it can
sometimes lag behind a bit, which might cause it to be replaced
with the regular one when performing a system update.

To work around the issue, perform the update in two steps: first
with all repositories enabled but excluding the kernel packages,
then only the kernel packages but with all repositories except
fedora-rawhide-kernel-nodebug disabled.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>