guests: Figure out OS-dependent paths automatically
When configuring a guest, we need to know the absolute path for
some commands and files; unfortunately, while most of them are
the same regardless of the OS, some of them don't.

Right now we hardcode such paths in the inventory, but it would
be better if we didn't have to do that: this commit introduces
a way to let Ansible figure out the paths for itself instead.

We're going to use it for more paths later on.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Pavel Hrdina <phrdina@redhat.com>