guests: Create ccache symlink farm
ccache can detect the name it's been called with and
transparently invoke the compiler if it matches. The Fedora
package uses this trick to enable ccache out of the box, but
other operating systems and distributions leave the task up
to the user.

Replicate the setup in the user's ~/.ccache directory so
that it works across all guests. With the link farm in
place, setting $CC in the user's shell profile is no longer
necessary.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Pavel Hrdina <phrdina@redhat.com>