Revert "projects: Drop explicit --enable-gtk-doc"
Turns out using --enable-gtk-doc *is* required, as otherwise
the *-rpm jobs for projects using gtk-doc fail with

  *** gtk-doc must be installed and enabled in order to make dist

Interestingly, libosinfo doesn't seem to be affected by this
issue despite it, too, using gtk-doc. Fixing libvirt-sandbox
and libvirt-glib so they don't need this either is left as an
exercise for the reader.

This reverts commit 14f444fec9ca7aa670ce7a8db62ba44cad2df6fd.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>