guests: Set MAKEFLAGS for out-of-the-box parallel make
We know all jobs are compatible with parallel make, and
we also know exactly how many vCPUs guests have, since we
created them in the first place: that allows us to default
to parallel make instead of making it opt-in.

The main advantage of doing this is that it will make life
more convenient for developers running one-off jobs or
debugging interactively.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>