guests: Remove distinction between files and templates
The main difference between files and templates is that the
former can can be uploaded as-is, while the latter has to
go through Jinja2 first.

Since the overhead of template processing is negligible, and
there's a fair chance we will need to start using templating
for files that currently don't need it or viceversa, let's
get rid of the distinction altogether.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>