jobs: Change git URL to fast external mirrors
We had our own local copy of all projects synchronized by cron on
the host where we have the CI VMs.  This was to save the traffic from
libvirt.org repositories and to make the cloning for our Jenkins jobs
faster.

We might move our VMs into a cloud in future we would not be able to
have any local copy so this changes the git URL to use fast external
mirrors.

Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>
Signed-off-by: Pavel Hrdina <phrdina@redhat.com>