lcitool: Pass base and playbook_base to Ansible
We want to get rid of relative paths in playbooks and
tasks, and in order to do that we have to provide Ansible
with some more information.

base is the directory where lcitool lives, and
playbook_base is the directory where a playbook should
look for its private resources: they match for the time
being, but that will no longer be the case very shortly.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Erik Skultety <eskultet@redhat.com>