guests: Drop install_config from inventory
The information is mostly duplicated and can be easily
inferred in a programmatic manner, so storing it in the
inventory is far from the cleanest solution.

As a side-effect, we reintroduce the error message that
was supposed to be displayed when attempting to install
a FreeBSD guest but was lost in the Python rewrite.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Erik Skultety <eskultet@redhat.com>