lcitool: Update OS when building Docker images
The last build of the Fedora Rawhide Docker image failed
with

  Transaction check error:
    file /usr/lib64/libgdbm_compat.so.4.0.0 from install
    of gdbm-libs-1:1.16-1.fc29.x86_64 conflicts with file
    from package gdbm-1:1.14.1-3.fc28.x86_64

caused by the gdbm package having recently been split.

Especially with fast-moving distribution such as Fedora
Rawhide and Debian Sid, we can't rely on the base OS not
changing over time, so merely installing the packages we
need on top of that is not enough.

Additionally, updating the OS brings the Docker images
closer to the guests running on CentOS CI, which are
updated daily.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Pavel Hrdina <phrdina@redhat.com>