jobs: Remove archive_format from defaults
Instead of defining a default and overriding it on a
case-by-case basis, always define archive_format at the
project level.

This adds a bit of duplication, but it's consistent
with how we define other metadata for projects; more
importantly, the current behavior would be extremely
difficult to implement in Ansible without introducing
extra boilerplate or deviating significantly from the
Jenkins version, neither of which we really want to do.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Erik Skultety <eskultet@redhat.com>