jobs: ensure rpmbuild purges $RPM-ROOT/BUILD/$PACKAGE
When using 'rpmbuild --rebuild', $RPM-ROOT/BUILD/$PACKAGE is
automatically deleted on completion, but when using 'rpmbuild -ta' it
is not deleted. We need to pass --clean to get the desired behaviour.

This was not a visible problem in the past because "git clean -fdx"
would purge the directory, but since some packages use %autosetup now,
the directories may contain a git repo which causes "git clean" to skip
deletion. For example with libvirt:

  $ su - jenkins
  $ cd libvirt-master
  $ git clean -fdx
  Skipping repository build/rpmbuild/BUILD/libvirt-4.7.0

Reviewed-by: Andrea Bolognani <abologna@redhat.com>
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>