Don't use "branch" in paths and job names
We'll soon to make it possible to build arbitrary branches
from arbitrary git repositories with lcitool: sticking with
the current scheme for on-disk storage would mean that we
would create a separate git clone for each branch that was
ever built this way, which is clearly extremely wasteful
not to mention slow. To prevent that from happening we drop
the branch name from the directory name, which makes it
possible to have a single git clone per project and allows
to quickly and cheaply switching repositories and branches.

After dropping the branch name from the directory name,
there is very little reason to keep it in the job name:
on the CentOS CI environment we only ever build master,
and with lcitool the users themselves provide the git
configuration including the branch name, so it doesn't
really provide any useful information in either case.
Let's drop it from there as well.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Erik Skultety <eskultet@redhat.com>