projects: Fix "git_url"
The "upstream" name was used in the initial version
of the patches, but during review it was decided to
change it to "default": regrettably, while the
"git_urls" dictionary was updated accordingly before
pushing, references to it in the projects/ directory
were not, which means they are now referring to
non-existing dictionary entries. Fix them.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>