guests: Install Python 3 during bootstrap
Almost all our target operating systems have a version
of Python 3 which is recent enough (>= 3.5) to be used
by Ansible available.

In fact, almost all of them have it pre-installed, which
makes installing Python 2 in addition to it just to use
Ansible, when Ansible itself would be absolutely happy
running on Python 3, quite silly.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>