guests: Store OS details as strings
When parsing something like

  os_version: 9

YAML will, somewhat reasonably, decide that os_version is
a number rather than a string.

This would be okay if not for the fact that in-development
versions of operating systems are usually assigned a
codename rather than a version number (Debian sid, Fedora
Rawhide, FreeBSD -CURRENT) and so we will always need to
convert the version to string before comparison to account
for that.

Quote all details about the OS to make sure they're stored
as strings to begin with.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>