mappings: extend mapping to allow per-arch entries
The architecture specific matches are considered after
all the non-architcture matches.

The dockerfile generator is updated to apply arch filtering based
on the host arch.

The Ansible playbook is not actually implementing support for
non-x86_64 architectures in this commit. It is just hardcoding
x86_64, which is enough to ensure the changes in the mappings.yml
file are a no-op initially.

Reviewed-by: Andrea Bolognani <abologna@redhat.com>
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>