lcitool: support generating cross compiler dockerfiles
Debian's filesystem layout has a nice advantage over Fedora which is
that it can install non-native packages in the main root filesystem. It
is thus possible to prepare an x86_64 filesystem containing -dev packages
for a foreign architecture, along with a GCC cross compiler.

QEMU has used this technique to facilitate developer build testing of
non-x86 architectures, since few people have access to physical
hardware for most of these architectures. For the same reason it would
be helpful to libvirt developers.

This patch extends the 'dockerfile' command to 'lcitool' so that it
accepts a '-x $ARCH' argument.

  $ lcitool dockerfile -x s390x libvirt-debian-9 libvirt

This is only valid when using Debian OS versions.

Reviewed-by: Andrea Bolognani <abologna@redhat.com>
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>