lcitool: include gdb, strace, lsof & netstat by default
When debugging why tests fail it is important to have commonly
used tools like gdb, strace, lsof and netstat available.

Note that although the entry for FreeBSD is blank for strace
and net-tools, the truss (strace equiv) and netstat tools will
be present as they are part of the FreeBSD base install.

Reviewed-by: Andrea Bolognani <abologna@redhat.com>
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>