guests: Install ncurses for libvirt
As of libvirt commit c98de2173e95, we use pkg-config to obtain
CFLAGS and LIBS for readline whenever possible. However, since
the readline.pc file contains a Requires.private for a module
that's shipped with ncurses (the actual module varies based on
the platform) we will end up falling back to the old library
detection if said library is not installed on the system.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>