lcitool: Don't warn when using package manager directly
We only do this when performing operations that the
corresponding Ansible module doesn't support, so we know
what we're doing and don't want warnings to show up.

Note that while only the dnf and yum modules complain at
the moment, we might as well use warn=no everywhere so that
we're already covered in case in the future the pkgng module
starts detecting this as well.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>