guests: Update mappings for sheepdog
Debian has removed sheepdog from the distribution, and at this
point Debian 9 is now the only supported Debian release that
still includes it.

Aside from that lone Debian release, Fedora and Ubuntu are the
only operating systems still including sheepdog, which means
that doing so is more the exception than the rule. Rewrite the
mapping accordingly.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>