guests: Use OpenJDK 11 where possible
Debian 10 doesn't include OpenJDK 8, which would be a problem
if not for the fact that thankfully the Jenkins agent, which
is the only thing we need Java for, seems to work perfectly
fine with OpenJDK 11 too.

Given this fact, we can move to OpenJDK 11 pretty much across
the board and stick to OpenJDK 8 for the few platforms that
only include the older release.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>