lcitool: Decouple Python interpreters for Ansible and builds
We need to keep using Python 2 for Ansible, because Python 3 is
only supported by the dnf module and on CentOS 7 we have to use
the yum module instead; at the same time, all libvirt projects
have now dropped Python 2 support so we need to make sure we use
Python 3 for building them.

Decouple the Python versions used for the two purposes so that
we can keep everything working.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>