lcitool: Add 'install=' to the extra_arg
Add "install=" unconditionally to the extra_arg parameter passed to
virt-install. It's needed as the virt-install present on CentOS CI host
does not use the correct kernel URL argument when installing openSUSE.

This does not cause issues on other installations as the other distros
will just ignore the argument passed.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Andrea Bolognani <abologna@redhat.com>