lcitool: Generate the unattended script file
Let's always generate the unattended script file as, by doing this, we
can treat the files we have as templates, being able to change them
accordingly to whatever is needed by specific distros.

It's important to note that we *must* be careful and keep generating
those files using their "expected" filename, preferably using the same
name of the template, as some of the distros require that.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Reviewed-by: Andrea Bolognani <abologna@redhat.com>