guests: Rename Ubuntu targets
Back when they were introduced, we chose to only include the part
of the version number preceding the dot with the rationale that
doing so introduced no ambiguity since, while there are two Ubuntu
releases per year, we only support LTS releases and those happen
once every two years.

When introducing openSUSE support, however, we could not adopt the
same strategy because openSUSE Leap 15.0 and openSUSe Leap 15.1
are, at least in some regards, just as distinct as Ubuntu 16.04 and
Ubuntu 18.04 are, despite sharing the same major version number;
we've thus decided to encode the full version number, dropping the
dot to avoid issues connected to the fact that we use the target
name as host name for guests.

Let's retroactively apply this new convention to existing Ubuntu
targets in order to maintain consistency.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>