jenkins: Don't build on CentOS 8 just yet
We're currently waiting to see whether we can get more capacity
for our CI environment, and in the meantime we've decided not do
create additional builders on the existing machine.

In order to be able to deploy other configuration changes without
enabling CentOS 8 jobs at the same time, we need to temporarily
drop all mentions of the target from the Jenkins configuration:
once we've figured out where to create the CentOS 8 builder, we
can simply revert this patch.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>