guests: optionally wait for installation completion
The install command is hardcoded to put the virt-install command into
the background. It is useful to be able to wait for completion, so you
know when you can go onto the package update step of setup, without
having to check back to see if the guest has finished installing.

Using --noautoconsole is preferred syntax for virt-install, avoiding
a warning about use of "--wait 0".

Reviewed-by: Erik Skultety <eskultet@redhat.com>
Tested-by: Erik Skultety <eskultet@redhat.com>
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>