lcitool: Fix pip use on CentOS 7
The 'pip' Ansible module doesn't work correctly when there is a
mismatch between the Python version we're adding packages to and
the one used to run Ansible itself: this is the case on CentOS 7,
where we have Python 3 for the former but Python 2 for the latter.

Call pip3 manually using the 'command' module instead to work
around this issue.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>