lcitool: Support MinGW cross-build Dockerfiles on Fedora
Up until this point, MinGW has been treated as a very special
case, even having its own sub-projects.

While this was perfectly reasonable back when lcitool was
introduced, now that we support cross-building as a first-class
concept in the Dockerfile generator it makes little sense for
MinGW builds not to make use of it.

Note that, as of this commit, MinGW is still considered somewhat
special: in particular, only the Dockerfile generator can treat
it like any other architecture, and everything else still needs
to use the sub-projects. In the longer run we'll want to expand
the first-class cross-building support to the rest of lcitool;
this is merely another step in that direction.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>