lcitool: Install dpkg-dev when doing cross-builds on Debian
When invoked as $triplet-pkg-config, pkg-config is smart enough
to search the paths corresponding to the target architecture; on
Debian, however, this currently only works when dpkg-dev is
installed, which by default it is not.

Work around the issue on our side by installing dpkg-dev
explicitly in our cross-building container images.

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=952526

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>