lcitool: Don't specify --target in $CONFIGURE_OPTS
If --target is not specified, the value of --host will be used,
so having both is unnecessary; --target is mostly intended for
building toolchains, which we're not doing.

In fact, we're already not using --target in either the Jenkins
and Ansible build scripts, thus proving that we don't need it.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>