lcitool: Include some paths in the generated Dockerfiles
These paths are a reflections of the contents of the container
image, so it makes sense to expose them in the container's
environment.

This will allow the GitLab CI integration for both libvirt and
libosinfo to stop setting these values themselves and start
relying on them being already present in the environment instead.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>