lcitool: Fix $PYTHON value in environment
Starting with commit

  commit d76f41ece09b9aaf492553fe466c30a22c78e509
  Author: Andrea Bolognani <abologna@redhat.com>
  Date:   Fri Mar 27 17:21:26 2020 +0100

    lcitool: Include paths in the inventory

    Figuring them out at runtime is neat, but in practice they don't
    change frequently enough for it to be necessary; more importantly,
    including them in the inventory means we can use them in the
    Dockerfile generator in addition to the Ansible playbooks.

    Signed-off-by: Andrea Bolognani <abologna@redhat.com>
    Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>

we have hardcoded the path of the Python interpreter to be used for
builds in the inventory; an unforeseen consequence of this change is
that Python builds on CentOS 7 have started failing. The reason for
that is simple, but not immediately obvious.

While on most target platforms we have a single Python interpreter
installed, on CentOS 7 we still need two: Python 3 works fine when
it comes to building projects, but it's missing a bunch of modules
that Ansible needs to run, so we're stuck with Python 2 for Ansible
use. This has worked fine for us so far.

As part of the change above, however, due to a naming clash we have
ended up in a situation where the bootstrap task, which installs a
Python suitable for Ansible, overrides the Python intended for builds
as set in the inventory. Of course this would result in breakages
only on CentOS 7, where the two Pythons differ.

Fix this by removing the runtime detection from the bootstrap task
and storing the relevant information in the inventory.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>