guests: users: Discard the .bash_logout skeleton file
As part of the users tasks, we remove the .profile file after we create
our own .bash_profile. As part of that effort, let's remove the .bash_logout
skeleton file as well, the reason being gitlab's runner agent which
has a hard time initializing a build environment on Debian where the
system populates the .bash_logout with console cleaning code which
messes up the runner and causes the build to fail instantly [1].
Since Debian is the only distro actually populating the file and the
machines are assumed to be accessed over SSH, we can safely drop it.

[1] https://gitlab.com/gitlab-org/gitlab-runner/issues/4449

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Reviewed-by: Andrea Bolognani <abologna@redhat.com>