vars: Add python3-wheel to the base package dependencies
On platforms where we need to install meson from pip, one will very
likely see something similar to this when building a container from the
generated Dockerfile:

    Collecting meson==0.49.0
      Downloading <url>/meson-0.49.0.tar.gz (1.3MB)
        100% |################################| 1.3MB 874kB/s
    Building wheels for collected packages: meson
      Running setup.py bdist_wheel for meson ... error
      error: invalid command 'bdist_wheel'
      ----------------------------------------
      Failed building wheel for meson
      Running setup.py clean for meson
    Failed to build meson

Pip is missing the 'wheel' package necessary to build a wheel from
sources, if it fails to do that, it falls back to the good old:
    $ setup.py install meson
which succeeds and no harm was done. However, seeing an error in the
log always raises eyebrows, so let's fix that very simply by installing
the 'wheel' package which is available on all supported platforms.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Reviewed-by: Andrea Bolognani <abologna@redhat.com>