guests: Add Ubuntu 20.04
Note that 'lcitool install' is not currently supported for this
platform, because Ubuntu switched to a new installer which
virt-install doesn't know about and we have no unattended
installation support for; the old installer is still available,
but it now lives in a different path and that means virt-install
doesn't know how to fetch it.

Once virt-install has been updated to support either form of
unattended installation for Ubuntu 20.04 we can update lcitool
accordingly, but for now this will do: we don't have any Ubuntu
VMs on CentOS CI anyway, and building Ubuntu 20.04 containers for
use on GitLab CI already works perfectly.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>