guests: Add cloud-init support
This patch adds support for generating images that can be used as
generic template images in a cloud environment by utilizing the
cloud-init service.
This feature is exposed as an option in the config.yaml which is
disabled by default. At the same time, this is treated as a new
internal project that filtered out of the list of projects we expose.
Unlike Linux-like distros where we simply create package mappings,
FreeBSD doesn't fully support cloud-init which is still a rolling
project and such has to be installed from a public git forge
repository.

Signed-off-by: Erik Skultety <eskultet@redhat.com>